rename: UserFixture 클래스 fixture 패키지로 이동 #4
